+++ src/main/java/org/wing4j/rrd/net/protocol/RoundRobinIncreaseProtocolV1.java
@@ -85,7 +85,7 @@ public class RoundRobinIncreaseProtocolV1 extends BaseRoundRobinProtocol {
         //版本号
         //报文类型
         //应答编码
-        this.code = buffer.getShort();
+        this.code = buffer.getInt();
         //应答描述
         this.desc = get(buffer);
         //表名长度
