+++ src/main/java/org/wing4j/rrd/net/protocol/RoundRobinTableMetadataProtocolV1.java
@@ -94,7 +94,7 @@ public class RoundRobinTableMetadataProtocolV1 extends BaseRoundRobinProtocol {
         //版本号
         //报文类型
         //应答编码
-        this.code = buffer.getShort();
+        this.code = buffer.getInt();
         //应答描述
         this.desc = get(buffer);
         //表名长度
